refactor(gateways): hoist findModelRecord into shared http helpers
Both vercel and openrouter adapters carried byte-identical
findModelRecord copies (same body, same JSDoc, same @kuralPure
@kuralHelper tags) — the kural duplicates audit flagged them at 100%
similarity. The function reads the OpenAI-compatible /v1/models
`data` array shape, which is owned by gateways/http.ts alongside the
structurally identical readFirstEndpoint, not by either gateway.

Move findModelRecord into http.ts, retag @kuralHelper -> @kuralUtil
to match readFirstEndpoint's role, and import it from both adapters.
